Bug Fix
The program would hang when trying to access a calibration file not on
the LRC's server. This has now been fixed and tested by disconnecting a
machine from the network. The problem was due to trying to close a file
that had never been open and an if True statement.