Extending Logging
Extending the logging of the Daysimeter Client to include a
daysimeter.log file which tracks program flow. This, in combination with
the error.log may be used to better diagnose where things aren't
working. The logging added is not complete (i.e. not every little task
is logged), but should be sufficient to diagnose problems.